L1 Markdown transcript, improved May 23, 2023
Lecture 1 transcript (Markdown version). Dave completed his first round of edits May 23, 2023 (04:20 UTC).
NOTES: (1) As of 5/23, this is the only well-edited version of the first lecture. The .txt version in this folder is NOT up to date.
(2) See the lists of Subject and People in the accompanying Supplements folder. They are in sync with this present edition.